pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neura_ai"
version = "0.1.5"
description = "Assistente IA com memória e voz"
dependencies = [
    "ollama",
    "SpeechRecognition",
    "pyttsx3",
    "pyaudio",
    "Pillow",
    "requests"
]

[tool.setuptools.packages.find]
include = ["neura_ai*"]  # Isso garante que apenas o código da lib seja instalado